Split out the atlas code from Theme.cpp into UIAtlas.cpp
