+++ UI/Theme.h
@@ -27,8 +27,4 @@ void ReloadAllThemeInfo();
 
 std::vector<std::string> GetThemeInfoNames();
 void UpdateTheme();
-Atlas *GetFontAtlas();
-Atlas *GetUIAtlas();
 UI::Theme *GetTheme();
-
-AtlasData AtlasProvider(Draw::DrawContext *draw, AtlasChoice atlas);
