+++ UI/UIAtlas.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "Common/UI/Context.h"
+#include "Common/Render/TextureAtlas.h"
+
+const Atlas *GetFontAtlas();
+Atlas *GetUIAtlas();
+AtlasData AtlasProvider(Draw::DrawContext *draw, AtlasChoice atlas);
